requirements.txt
albumentations==1.3.1
lightning==2.1.2
clearml==1.13.2
pandas==2.1.4
tensorboard==2.15.1
tensorboardX==2.6.2.2
segmentation-models-pytorch==0.3.3
opencv-python==4.8.1.78
pydantic==2.5.2
omegaconf==2.3.0
torchvision==0.16.1
scikit-image==0.22.0
wldhx.yadisk-direct==0.0.6
onnx==1.15.0
onnxruntime==1.16.3
jupyter==1.0.0
dvc[ssh]==3.33.3